requirements.txt
pandas
sqlalchemy
psycopg2-binary
panel
python-dotenv
jupyter_bokeh